move getTreeKey methods to separate file
