+++ src/js/verkleFFIBindings/index.ts
@@ -1,106 +1,13 @@
 import { Context, zeroCommitment } from '../../../dist/cjs/wasm/rust_verkle_wasm'
 
+import { getTreeKeyJs } from './getTreeKey'
+
 // This is a 32 byte serialized field element
 type Scalar = Uint8Array
 
 // This is a 64 byte serialized point.
 // It is 64 bytes because the point is serialized in uncompressed format.
 type Commitment = Uint8Array
-
-export function getTreeKeyJs(
-    context: Context,
-    address: Uint8Array,
-    treeIndex: Uint8Array,
-    subIndex: number,
-): Uint8Array {
-    const serializedCommitment = getTreeKeyHashJs(context, address, treeIndex)
-
-    // Replace the last byte with the subIndex
-    serializedCommitment[serializedCommitment.length - 1] = subIndex
-    return serializedCommitment
-}
-
-// Computes the hash of the address, treeIndex and subIndex
-export function getTreeKeyHashJs(
-    context: Context,
-    address: Uint8Array,
-    treeIndex: Uint8Array,
-): Uint8Array {
-    if (address.length !== 32) {
-        throw new Error('Address must be 32 bytes')
-    }
-    if (treeIndex.length !== 32) {
-        throw new Error('Tree index must be 32 bytes')
-    }
-
-    /*
-        Here is the function we wish to implement here:
-        def pedersen_hash(inp: bytes) -> bytes32:
-            assert len(inp) <= 255 * 16
-            # Interpret input as list of 128 bit (16 byte) integers
-            ext_input = inp + b"\0" * (255 * 16 - len(inp))
-            ints = [2 + 256 * len(inp)] + \
-                [int.from_bytes(ext_input[16 * i:16 * (i + 1)]) for i in range(255)]
-            return compute_commitment_root(ints).serialize()
-        */
-
-    const input = concatenateUint8Arrays(address, treeIndex)
-
-    // The input is chopped up into 16 byte chunks (u128 integers) but since the input has constant size
-    // Lets compute the number of 16 byte chunks needed and other constants.
-    const chunkSize = 16
-    //
-    //
-    // Address is a 32 byte array
-    // treeIndex is a 32 byte array
-    // In total, we have 64 bytes
-    // This means that len(inp) will always be 64.
-    //
-    // 64 /16 = 4 but notice that an extra 16 byte chunk is prepended to the input
-    // to encode the type of hash and the length. This is `[2 + 256 * len(inp)]`
-    // This means that the number of 16 byte chunks in total will be 5.
-
-    // As noted the first chunk will always be [2 + 256 * 64] because len(inp) is always 64.
-    const firstChunk = 2n + 256n * 64n
-
-    const chunks: Uint8Array[] = []
-    // TODO: This is a constant, so we don't need to compute it every time
-    // TODO: [2, 64, 0, 0 ,0,..0]
-    chunks.push(bigintToUint8ArrayLE(firstChunk, chunkSize))
-
-    for (let i = 0; i < input.length; i += chunkSize) {
-        const chunk = input.slice(i, i + chunkSize)
-        chunks.push(chunk)
-    }
-
-    const commitment = context.commitTo16ByteScalars(chunks)
-    const serializedCommitment = context.deprecateSerializeCommitment(commitment).reverse()
-    return serializedCommitment
-}
-
-function concatenateUint8Arrays(array1: Uint8Array, array2: Uint8Array): Uint8Array {
-    const concatenatedArray = new Uint8Array(array1.length + array2.length)
-
-    concatenatedArray.set(array1, 0)
-    concatenatedArray.set(array2, array1.length)
-
-    return concatenatedArray
-}
-
-function bigintToUint8ArrayLE(value: bigint, byteLength: number): Uint8Array {
-    // Allocate a buffer of the desired size
-    const buffer = new ArrayBuffer(byteLength)
-    const view = new DataView(buffer)
-
-    // Fill the buffer from the end, as BigInt is little-endian
-    for (let i = 0; i < byteLength; i++) {
-        // Extract byte by shifting right (i * 8) bits and mask off a byte
-        const byte = Number((value >> BigInt(i * 8)) & BigInt(0xff))
-        view.setUint8(i, byte)
-    }
-
-    return new Uint8Array(buffer)
-}
 
 export {
     Scalar,
@@ -109,4 +16,5 @@ export {
     // This is a function that returns a zero commitment
     // wasm_bindgen does not seem to allow returning constants
     zeroCommitment,
+    getTreeKeyJs,
 }
